update: 항목 검증 첨가

+++ src/checkitems/checkitems.service.ts
@@ -1,4 +1,9 @@
-import { Injectable, InternalServerErrorException } from '@nestjs/common';
+import {
+  BadRequestException,
+  Injectable,
+  InternalServerErrorException,
+  NotFoundException,
+} from '@nestjs/common';
 import { CreateCheckitemDto } from './dto/create-checkitem.dto';
 import { UpdateCheckitemDto } from './dto/update-checkitem.dto';
 import { CheckItem } from './entities/checkitem.entity';
@@ -7,7 +12,6 @@ import { Repository } from 'typeorm';
 
 @Injectable()
 export class CheckitemsService {
-  //private checkitems: CheckItem[] = []; // 임시 저장소
   constructor(
     @InjectRepository(CheckItem)
     private checkitemsRepository: Repository<CheckItem>,
@@ -48,13 +52,18 @@ export class CheckitemsService {
     id: number,
     updateCheckitemDto: UpdateCheckitemDto,
   ): Promise<CheckItem> {
+    const { checklistId, title, status, position, memberId } =
+      updateCheckitemDto;
     try {
-      //카드id 검증 필요
-      //멤버id 검증 필요
-      const checkitem = await this.checkitemsRepository.findOneBy({ id }); // ID로 항목 찾기
+      const checkitem = await this.checkitemsRepository.findOneBy({ id }); // ID로 체크리스트 찾기
       if (!checkitem) {
-        throw new Error('체크리스트 항목을 찾을 수 없습니다.'); // 에러 처리
+        throw new NotFoundException('체크리스트 항목을 찾을 수 없습니다.'); // 에러 처리
       }
+      //체크리스트 id 검증
+      if (checkitem.checkListId !== checklistId) {
+        throw new BadRequestException('체크리스트의 ID가 일치하지 않습니다.');
+      }
+      //멤버id 검증 필요
 
       Object.assign(checkitem, updateCheckitemDto);
       return this.checkitemsRepository.save(checkitem); // 업데이트된 항목 저장
@@ -64,14 +73,23 @@ export class CheckitemsService {
   }
 
   // 항목 삭제 메서드
-  async remove(id: number, checkListId: number): Promise<void> {
+  async remove(
+    id: number,
+    updateCheckitemDto: UpdateCheckitemDto,
+  ): Promise<void> {
+    const { checklistId } = updateCheckitemDto;
     try {
       const checkitem = await this.checkitemsRepository.findOneBy({
         id,
       });
 
       if (!checkitem) {
-        throw new Error('항목이 없어용.'); // 에러 처리
+        throw new NotFoundException('항목이 없어용.'); // 에러 처리
+      }
+
+      //체크리스트 id 검증
+      if (checkitem.checkListId !== checklistId) {
+        throw new BadRequestException('체크리스트의 ID가 일치하지 않습니다.');
       }
 
       await this.checkitemsRepository.remove(checkitem); // 항목 삭제
